third_party/WebKit/LayoutTests/http/tests/websocket/cookie-http-to-ws-expected.txt

Test how WebSocket handles cookies with cookie attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cookie is "WK-websocket-test-domain=1; WK-websocket-test-path=1"
PASS cookie is "WK-websocket-test-domain=1"
PASS successfullyParsed is true

TEST COMPLETE
